# specs/001-commercial-management-mvp/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Commercial Management MVP API
  version: 1.0.0
  description: REST API for Commercial Management MVP - Products, Customers, Stock, Sales (Quotes/Orders), and Dashboard
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
...
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [B2B, B2C]
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
                  total:
                    type: integer
    post:
      tags: [Customers]
      summary: Create customer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
  /customers/{id}:
    get:
      tags: [Customers]
      summary: Get customer by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'

  # Stock
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        type:
          type: string
          enum: [B2B, B2C]
        name:
          type: string
        email:
          type: string
          format: email
        status:
          type: string
          enum: [active, inactive, archived, blocked]
    CustomerCreate:
      type: object
      required: [type, name, email]
      properties:
        type:
          type: string
          enum: [B2B, B2C]
        name:
          type: string
        email:
          type: string
          format: email
        company_name:
          type: string
          description: Required if type is B2B
        first_name:
          type: string
          description: Required if type is B2C
        last_name:
          type: string
          description: Required if type is B2C
...